+++ Core/Native/ArchetypeExtensionsForNative.cs.meta
