+++ Core/Native/ComponentsCacheExtensionsForNative.cs.meta
