+++ Core/Native/NativeArchetypeWrapper.cs.meta
