+++ Core/Native/NativeCache.cs
@@ -1,5 +1,5 @@
 ﻿#if MORPEH_BURST
-namespace Morpeh.NativeCollections {
+namespace Morpeh.Native {
     using System.Runtime.CompilerServices;
     using Morpeh;
     using Unity.Collections.LowLevel.Unsafe;
