+++ Core/Native/NativeFastListWrapper.cs
@@ -1,5 +1,5 @@
 ﻿#if MORPEH_BURST
-namespace Morpeh.NativeCollections {
+namespace Morpeh.Native {
     using Unity.Collections.LowLevel.Unsafe;
 
     public struct NativeFastListWrapper<TNative> where TNative : unmanaged {
